+++ src/main/java/com/atlauncher/constants/Constants.java
@@ -61,6 +61,7 @@ public class Constants {
     public static String API_BASE_URL = BASE_LAUNCHER_PROTOCOL + "api." + BASE_LAUNCHER_DOMAIN + "/v1/launcher/";
     public static String GRAPHQL_ENDPOINT = BASE_LAUNCHER_PROTOCOL + "api." + BASE_LAUNCHER_DOMAIN + "/v2/graphql";
     public static String API_HOST = "api." + BASE_LAUNCHER_DOMAIN;
+    public static String ANALYTICS_BASE_URL = BASE_LAUNCHER_PROTOCOL + "analytics." + BASE_LAUNCHER_DOMAIN;
     public static String PASTE_CHECK_URL = BASE_LAUNCHER_PROTOCOL + "paste." + BASE_LAUNCHER_DOMAIN;
     public static String PASTE_HOST = "paste." + BASE_LAUNCHER_DOMAIN;
     public static String SERVERS_LIST_PACK = BASE_LAUNCHER_PROTOCOL + BASE_LAUNCHER_DOMAIN + "/servers/list/pack";